+++ lib/utils/src/public_api.ts
@@ -1,3 +1,0 @@
-export { createEntityCollections } from './entity/create-entity-collections';
-export { NgxsDataMutablePipe } from './mutable/ngxs-data-mutable.pipe';
-export { NgxsDataUtilsModule } from './ngxs-data-utils.module';